Add chat service
