+++ service/chat/config/config.json
@@ -8,7 +8,7 @@
       "InternalListenAddress" : ":19802"
    },
    "DBConf" : {
-      "DBName" : "momo",
+      "DBName" : "chat",
       "Host" : "192.168.1.103",
       "Port" : 3306,
       "Password" : "koudai123456",
