--- service/chat/handlers/handlers.go
@@ -1,10 +1,0 @@
-package handlers
-
-import (
-	"fxlibraries/errors"
-	"fxlibraries/httpserver"
-)
-
-func Test(req *httpserver.Request) *httpserver.Response {
-	return httpserver.NewResponseWithError(errors.InternalServerError)
-}
